Force the default collection withdrawal delay to be always of type integer

+++ includes/functions.php
@@ -1581,10 +1581,10 @@ function incassoos_get_custom_email_closing( $default = '' ) {
  * @uses apply_filters() Calls 'incassoos_get_default_collection_withdrawal_delay'
  *
  * @param mixed $default Optional. Default return value.
- * @return string Default collection withdrawal delay
+ * @return int Default collection withdrawal delay
  */
 function incassoos_get_default_collection_withdrawal_delay( $default = 5 ) {
-	return apply_filters( 'incassoos_get_default_collection_withdrawal_delay', get_option( '_incassoos_default_collection_withdrawal_delay', $default ) );
+	return (int) apply_filters( 'incassoos_get_default_collection_withdrawal_delay', get_option( '_incassoos_default_collection_withdrawal_delay', $default ) );
 }
 
 /**
